pyproject.toml
[project]
name = "nonebot_plugin_litestore"
version = "0.1.0"
description = "A Lighter LocalStore For NoneBot2"
authors = [
    {name = "kanbereina", email = "kano.2525@qq.com"},
]
dependencies = ["nonebot2>=2.4.1", "typing-extensions>=4.12.2"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
